requirements.txt
flask
flask-sqlalchemy
mysql-connector-python
flask-bcrypt
pyjwt
flask-jwt-extended